+++ test/unittest/TClosure.cpp
@@ -9,11 +9,21 @@ void Foo0() {
 void Foo1(int x) {
 }
 
+void Foo2(int x, float y) {
+}
+
+void Foo3(int x, float y, double z) {
+}
+
 class Obj {
  public:
   void Foo0() {
   }
-  void Foo1(int i) {
+  void Foo1(int x) {
+  }
+  void Foo2(int x, float y) {
+  }
+  void Foo3(int x, float y, double z) {
   }
 }; 
 
@@ -32,6 +42,22 @@ TEST(Closure, Test0) {
   delete c;
   c = NULL;
 
+  c = NewClosure(&Foo2, 1, float(2));
+  ASSERT_TRUE(c != NULL);
+  c->Run();
+  delete c;
+  c = NULL;
+
+  c = NewClosure(&Foo3, 1, float(2), double(3));
+  ASSERT_TRUE(c != NULL);
+  c->Run();
+  delete c;
+  c = NULL;
+}
+
+TEST(Closure, Test1) {
+  ClosureInterface* c;
+
   Obj obj;
   c = NewClosure(&obj, &Obj::Foo0);
   ASSERT_TRUE(c != NULL);
@@ -40,6 +66,18 @@ TEST(Closure, Test0) {
   c = NULL;
 
   c = NewClosure(&obj, &Obj::Foo1, 1);
+  ASSERT_TRUE(c != NULL);
+  c->Run();
+  delete c;
+  c = NULL;
+
+  c = NewClosure(&obj, &Obj::Foo2, 1, float(2));
+  ASSERT_TRUE(c != NULL);
+  c->Run();
+  delete c;
+  c = NULL;
+
+  c = NewClosure(&obj, &Obj::Foo3, 1, float(2), double(3));
   ASSERT_TRUE(c != NULL);
   c->Run();
   delete c;